+++ distribs/debian_wheezy.sh
@@ -54,7 +54,7 @@ debian_wheezy_dep_install()
 	export DEBIAN_FRONTEND=noninteractive
 	apt-get -q -y --force-yes install build-essential curl subversion git-core checkinstall libcxxtools-dev yasm scons libboost-dev zlib1g-dev unzip \
 		apache2.2-common mysql-server php5-dev php5-mysql php5-sqlite php5-imagick php-pear php5-curl php5-gd libapache2-mod-php5 \
-		libmp3lame-dev libaacplus-dev libopencore-amrnb-dev libopencore-amrwb-dev libvpx-dev libtheora-dev librtmp-dev libfaac-dev libfaad-dev libmodplug-dev libgsm1-dev libopenjpeg-dev libxvidcore-dev libschroedinger-dev libspeex-dev libvorbis-dev libass-dev libtwolame-dev \
+		libmp3lame-dev libaacplus-dev libopencore-amrnb-dev libopencore-amrwb-dev libtheora-dev librtmp-dev libfaac-dev libfaad-dev libmodplug-dev libgsm1-dev libopenjpeg-dev libxvidcore-dev libschroedinger-dev libspeex-dev libvorbis-dev libass-dev libtwolame-dev \
 		flac vorbis-tools imagemagick pngnq optipng libjpeg-progs \
 		tesseract* xpdf catdoc unoconv poppler-utils texi2html \
 		2>> $LOG >> $LOG || return 1
@@ -66,6 +66,8 @@ debian_wheezy_dep_install()
 	debian_wheezy_libopus_install || return 1
 	
 	libfdk_aac_install || return 1
+	
+	libvpx_install || return 1
 	
 	flvtool_plus_install || return 1