+++ distribs/ubuntu_precise.sh
@@ -49,10 +49,10 @@ ubuntu_precise_dep_install()
 	ubuntu_precise_yasm_install || return 1
 
 	ubuntu_precise_libopus_install || return 1
-
-	ubuntu_precise_libvpx_install || return 1
 	
 	libfdk_aac_install || return 1
+
+	libvpx_install || return 1
 	
 	flvtool_plus_install || return 1
 
@@ -94,46 +94,6 @@ ubuntu_precise_yasm_install ()
 		checkinstall --pkgname=yasm --pkgversion "$VERSION+mediaspip" --backup=no --default 2>> $LOG >> $LOG || return 1
 		echo $(eval_gettext 'End $SOFT')
 	fi
-	echo
-}
-
-# Installation de libvpx
-# http://code.google.com/p/webm/
-ubuntu_precise_libvpx_install()
-{
-	export TEXTDOMAINDIR=$CURRENT/locale
-	export TEXTDOMAIN=mediaspip
-	SOFT="libvpx"
-	cd $SRC_INSTALL
-	VERSION="1.1.0"
-	LIBVPX=$(dpkg --status libvpx 2>> $LOG |awk '/^Version/ { print $2 }') 2>> $LOG >> $LOG
-	case "$LIBVPX" in
-		*$VERSION*)
-			echo $(eval_gettext 'Info a jour $SOFT $VERSION')
-			echo $(eval_gettext 'Info a jour $SOFT $VERSION') 2>> $LOG >> $LOG
-			;;
-		*)
-			echo $(eval_gettext 'Info debut $SOFT install $VERSION')
-			echo $(eval_gettext 'Info debut $SOFT install $VERSION') 2>> $LOG >> $LOG
-			if [ ! -e "$SRC_INSTALL"/libvpx-v1.1.0.tar.bz2 ];then
-				wget http://webm.googlecode.com/files/libvpx-v1.1.0.tar.bz2 2>> $LOG >> $LOG
-				tar xvjf libvpx-v1.1.0.tar.bz2 2>> $LOG >> $LOG
-			elif [ ! -d "$SRC_INSTALL"/libvpx-v1.1.0 ]; then
-				tar xvjf libvpx-v1.1.0.tar.bz2 2>> $LOG >> $LOG
-			fi
-			cd libvpx-v1.1.0
-			make -j $NO_OF_CPUCORES clean 2>> $LOG >> $LOG
-			echo $(eval_gettext "Info compilation configure")
-			./configure --enable-shared 2>> $LOG >> $LOG
-			echo $(eval_gettext "Info compilation make")
-			make -j $NO_OF_CPUCORES 2>> $LOG >> $LOG
-			echo $(eval_gettext "Info compilation install")
-			apt-get -y --force-yes remove libvpx 2>> $LOG >> $LOG
-			checkinstall --fstrans=no --install=yes --pkgname="libvpx" --pkgversion="$VERSION+mediaspip" --backup=no --default 2>> $LOG >> $LOG
-			echo $(eval_gettext 'End $SOFT')
-			;;
-	esac
-	ldconfig
 	echo
 }